relocating more functions since majority of the legacy code in officer_management_helper.py was relevant only to the import-export functions for officer lists
