+++ csss-site/src/about/views/github_position_mapping/update_saved_github_mappings.py
@@ -4,7 +4,7 @@ from django.shortcuts import render
 from querystring_parser import parser
 
 from about.models import OfficerEmailListAndPositionMapping
-from about.views.officer_position_and_github_mapping.officer_management_helper import TAB_STRING
+from about.views.Constants import TAB_STRING
 from about.views.position_mapping_helper import update_context, GITHUB_TEAM__ID_KEY, \
     GITHUB_TEAM_RELEVANT_PREVIOUS_TERM_KEY, validate_position_names_for_github_team, \
     OFFICER_EMAIL_LIST_AND_POSITION_MAPPING__DELETE_STATUS, GITHUB_MAPPING_SELECTED_OFFICER_POSITIONS, \
