+++ csss-site/src/about/views/officer_positions/input_new_officer_positions.py
@@ -4,7 +4,7 @@ from django.shortcuts import render
 from querystring_parser import parser
 
 from about.models import OfficerEmailListAndPositionMapping
-from about.views.officer_position_and_github_mapping.officer_management_helper import TAB_STRING
+from about.views.Constants import TAB_STRING
 from about.views.position_mapping_helper import update_context, validate_position_index, validate_position_name, \
     POSITION_INDEX_KEY, validate_elected_via_election_officer_status, validate_number_of_terms, \
     validate_starting_month, validate_github_access, validate_google_drive_access, \
