+++ csss-site/src/about/views/officer_positions/officer_positions.py
@@ -1,6 +1,6 @@
 from django.shortcuts import render
 
-from about.views.officer_position_and_github_mapping.officer_management_helper import TAB_STRING
+from about.views.Constants import TAB_STRING
 from about.views.position_mapping_helper import update_context
 from csss.views.context_creation.create_authenticated_contexts import create_context_for_updating_position_mappings
 from csss.views.views import ERROR_MESSAGES_KEY
